Wrapping the search's methods into search module
