+++ test/search_handler_test.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+"""Search handler test."""
